backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pymupdf
pillow
sentence-transformers
qdrant-client
groq
python-dotenv
httpx
aiofiles